% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapt.R
\name{scrapt}
\alias{scrapt}
\title{Start SCRAPT shiny server}
\usage{
scrapt(idata = NULL, threads = NULL)
}
\arguments{
\item{threads}{Number of threads to launch SCRAPT}
}
\description{
Start SCRAPT shiny server
}
