% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{SCRAPT server}
\usage{
server(input, output, session, threads = 5)
}
\arguments{
\item{input}{Server input variables}

\item{output}{Server output variables}

\item{session}{Current session}

\item{threads}{Number of threads to launch SCRAPT}
}
\description{
SCRAPT server
}
